# perf: cache a map of packages